[clients] Move Sponsor
