+++ src/modules/sponsor/getEdition.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk'
-import { Sponsor } from '../../clients/sponsor'
 import { createIOContext } from '../../lib/clients'
+import { Sponsor } from '../../lib/clients/Sponsor'
 import { SessionManager } from '../../lib/session/SessionManager'
 import log from '../../logger'
 import { IOClientOptions } from '../utils'
